+++ src/config.ts
@@ -23,7 +23,11 @@ const log = logging.createLogger('config');
 
 export type LogLevelKey = 'trace' | 'debug' | 'info' | 'note' | 'warning' | 'error' | 'fatal';
 export type CMakeCommunicationMode = 'legacy' | 'serverApi' | 'fileApi' | 'automatic';
-export type StatusBarButtonVisibility = "default" | "compact" | "icon" | "hidden";
+export type StatusBarButtonVisibility = "visible" | "compact" | "icon" | "hidden";
+export type StatusBarStaticButtonVisibility = "visible" | "icon" | "hidden";
+export type StatusBarTextButtonVisibility = "visible" | "compact" | "hidden";
+export type StatusBarIconButtonVisibility = "visible" | "hidden";
+export type ProjectStatusButtonVisibility = "visible" | "hidden";
 export type TouchBarButtonVisibility = "default" | "hidden";
 export type UseCMakePresets = 'always' | 'never' | 'auto';
 
@@ -39,53 +43,72 @@ export interface TouchBarConfig {
     visibility: TouchBarButtonVisibility;
 }
 
-export interface AdvancedStatusBarConfig {
+export interface AdvancedStatusConfig {
+    configure?: {
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+    };
     configurePreset?: {
-        visibility?: StatusBarButtonVisibility;
-        length?: number;
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarButtonVisibility;
+        statusBarLength?: number;
     };
     buildPreset?: {
-        visibility?: StatusBarButtonVisibility;
-        length?: number;
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarButtonVisibility;
+        statusBarLength?: number;
     };
     testPreset?: {
-        visibility?: StatusBarButtonVisibility;
-        length?: number;
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarButtonVisibility;
+        statusBarLength?: number;
     };
     kit?: {
-        visibility?: StatusBarButtonVisibility;
-        length?: number;
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarButtonVisibility;
+        statusBarLength?: number;
     };
-    status?: {
-        visibility?: StatusBarButtonVisibility;
+    variantStatus?: {
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarButtonVisibility;
     };
-    workspace?: {
-        visibility?: StatusBarButtonVisibility;
+    folder?: {
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarButtonVisibility;
+        statusBarLength?: number;
     };
     buildTarget?: {
-        visibility?: StatusBarButtonVisibility;
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarTextButtonVisibility;
+        statusBarLength?: number;
     };
     build?: {
-        visibility?: StatusBarButtonVisibility;
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarStaticButtonVisibility;
     };
     launchTarget?: {
-        visibility?: StatusBarButtonVisibility;
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarTextButtonVisibility;
+        statusBarLength?: number;
     };
     debug?: {
-        visibility?: StatusBarButtonVisibility;
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarIconButtonVisibility;
     };
     launch?: {
-        visibility?: StatusBarButtonVisibility;
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
+        statusBarVisibility?: StatusBarIconButtonVisibility;
     };
     ctest?: {
+        projectStatusVisibility?: ProjectStatusButtonVisibility;
         color?: boolean;
-        visibility?: StatusBarButtonVisibility;
+        statusBarVisibility?: StatusBarButtonVisibility;
+        statusBarLength?: number;
     };
 }
 
-export interface StatusBarConfig {
-    advanced?: AdvancedStatusBarConfig;
-    visibility: StatusBarButtonVisibility;
+export interface StatusConfig {
+    advanced?: AdvancedStatusConfig;
+    statusBarVisibility: StatusBarButtonVisibility;
 }
 
 export interface ExtensionConfigurationSettings {
@@ -138,8 +161,8 @@ export interface ExtensionConfigurationSettings {
     loggingLevel: LogLevelKey;
     additionalKits: string[];
     touchbar: TouchBarConfig;
-    statusbar: StatusBarConfig;
-    useProjectStatusView: boolean;
+    showOptionsMovedNotification: boolean;
+    status: StatusConfig;
     useCMakePresets: UseCMakePresets;
     allowCommentsInPresetsFile: boolean;
     allowUnsupportedPresetsVersions: boolean;
@@ -451,12 +474,8 @@ export class ConfigurationReader implements vscode.Disposable {
         return this.configData.touchbar;
     }
 
-    get statusbar() {
-        return this.configData.statusbar;
-    }
-
-    get useProjectStatusView(): boolean {
-        return this.configData.useProjectStatusView;
+    get status() {
+        return this.configData.status;
     }
 
     get launchBehavior(): string {
@@ -517,8 +536,8 @@ export class ConfigurationReader implements vscode.Disposable {
         loggingLevel: new vscode.EventEmitter<LogLevelKey>(),
         additionalKits: new vscode.EventEmitter<string[]>(),
         touchbar: new vscode.EventEmitter<TouchBarConfig>(),
-        statusbar: new vscode.EventEmitter<StatusBarConfig>(),
-        useProjectStatusView: new vscode.EventEmitter<boolean>(),
+        showOptionsMovedNotification: new vscode.EventEmitter<boolean>(),
+        status: new vscode.EventEmitter<StatusConfig>(),
         useCMakePresets: new vscode.EventEmitter<UseCMakePresets>(),
         allowCommentsInPresetsFile: new vscode.EventEmitter<boolean>(),
         allowUnsupportedPresetsVersions: new vscode.EventEmitter<boolean>(),